notebooks/llama-index-property-graph-testing/requirements.txt
llama-index
llama-index-graph-stores-neo4j 
llama-index-vector-stores-qdrant
llama-index-llms-azure-openai
python-dotenv
llama-index-embeddings-azure-openai
pyvis